Setting up proper context switching. Switching in works, switching out doesn't

+++ kernel/src/include/mem/pmm.h
@@ -22,7 +22,7 @@ extern volatile struct limine_memmap_request memmap_request;
 #define PHYS_ZONE_COUNT 15
 #define FREE 0x1
 #define USED 0x2
-#define UNUSED 0xFFFFFFFFFFFFFFFF
+#define UNUSED (uint64_t) 0xFFFFFFFFFFFFFFFF
 
 #define BUDDY_HASH_TABLE_SIZE 300
